refactored header_parser_test to use TempDir
