+++ gcp_variant_transforms/testing/temp_dir.py
